requirements.txt
boto3
Flask-Caching
flask
pyyaml
rgwadmin